refactor dynamic buffer includes
