+++ include/boost/buffers.hpp
@@ -11,15 +11,15 @@
 #define BOOST_BUFFERS_HPP
 
 #include <boost/buffers/buffer.hpp>
-#include <boost/buffers/copy.hpp>
-#include <boost/buffers/circular_buffer.hpp>
 #include <boost/buffers/buffer_pair.hpp>
+#include <boost/buffers/circular_buffer.hpp>
+#include <boost/buffers/copy.hpp>
+#include <boost/buffers/dynamic_buffer.hpp>
 #include <boost/buffers/flat_buffer.hpp>
 #include <boost/buffers/front.hpp>
 #include <boost/buffers/make_buffer.hpp>
 #include <boost/buffers/range.hpp>
-#include <boost/buffers/string_buffer.hpp>
 #include <boost/buffers/slice.hpp>
-#include <boost/buffers/type_traits.hpp>
+#include <boost/buffers/string_buffer.hpp>
 
 #endif
