+++ include/boost/buffers/dynamic_buffer.hpp
@@ -7,17 +7,56 @@
 // Official repository: https://github.com/cppalliance/buffers
 //
 
-#ifndef BOOST_BUFFERS_ANY_DYNAMIC_BUFFER_HPP
-#define BOOST_BUFFERS_ANY_DYNAMIC_BUFFER_HPP
+#ifndef BOOST_BUFFERS_DYNAMIC_BUFFER_HPP
+#define BOOST_BUFFERS_DYNAMIC_BUFFER_HPP
 
 #include <boost/buffers/detail/config.hpp>
 #include <boost/buffers/buffer.hpp>
-#include <boost/buffers/type_traits.hpp>
 #include <boost/core/span.hpp>
 #include <cstdlib>
 
 namespace boost {
 namespace buffers {
+
+/** Determine if T is a DynamicBuffer
+*/
+template<
+    class T,
+    class = void>
+struct is_dynamic_buffer : std::false_type {};
+
+template<class T>
+struct is_dynamic_buffer<
+    T, void_t<decltype(
+        std::declval<std::size_t&>() =
+            std::declval<T const&>().size()
+        ,std::declval<std::size_t&>() =
+            std::declval<T const&>().max_size()
+        ,std::declval<std::size_t&>() =
+            std::declval<T const&>().capacity()
+        ,std::declval<T&>().commit(
+            std::declval<std::size_t>())
+        ,std::declval<T&>().consume(
+            std::declval<std::size_t>())
+    )
+    ,typename std::enable_if<
+        is_const_buffer_sequence<typename
+            T::const_buffers_type>::value
+        && is_mutable_buffer_sequence<typename
+            T::mutable_buffers_type>::value
+        >::type
+    ,typename std::enable_if<
+        std::is_same<decltype(
+            std::declval<T const&>().data()),
+            typename T::const_buffers_type>::value
+        && std::is_same<decltype(
+            std::declval<T&>().prepare(
+                std::declval<std::size_t>())),
+            typename T::mutable_buffers_type>::value
+        >::type
+    > > : std::true_type
+{
+};
 
 /** An abstract, type-erased dynamic buffer.
 */
