+++ include/boost/buffers/range.hpp
@@ -12,7 +12,6 @@
 
 #include <boost/buffers/detail/config.hpp>
 #include <boost/buffers/buffer.hpp>
-#include <boost/buffers/type_traits.hpp>
 #include <type_traits>
 
 namespace boost {
