+++ include/boost/buffers/slice.hpp
@@ -13,7 +13,6 @@
 #include <boost/buffers/detail/config.hpp>
 #include <boost/buffers/buffer.hpp>
 #include <boost/buffers/range.hpp>
-#include <boost/buffers/type_traits.hpp>
 #include <boost/assert.hpp>
 #include <iterator>
 #include <type_traits>
