--- include/boost/buffers/tag_invoke.hpp
@@ -1,23 +1,0 @@
-//
-// Copyright (c) 2023 Vinnie Falco (vinnie.falco@gmail.com)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-// Official repository: https://github.com/cppalliance/buffers
-//
-
-#ifndef BOOST_BUFFERS_TAG_INVOKE_HPP
-#define BOOST_BUFFERS_TAG_INVOKE_HPP
-
-#include <boost/buffers/detail/config.hpp>
-#include <boost/buffers/buffer.hpp>
-#include <utility>
-
-namespace boost {
-namespace buffers {
-
-} // buffers
-} // boost
-
-#endif
