+++ test/unit/Jamfile
@@ -22,21 +22,19 @@ project
 
 local SOURCES =
     test_buffers.hpp
-    any_dynamic_buffer.cpp
     asio.cpp
     buffer.cpp
     buffer_pair.cpp
     buffers.cpp
     circular_buffer.cpp
     copy.cpp
+    dynamic_buffer.cpp
     flat_buffer.cpp
     front.cpp
     make_buffer.cpp
     range.cpp
     slice.cpp
     string_buffer.cpp
-    tag_invoke.cpp
-    type_traits.cpp
     ;
 
 for local f in $(SOURCES)
