+++ test/unit/flat_buffer.cpp
@@ -10,7 +10,7 @@
 // Test that header file is self-contained.
 #include <boost/buffers/flat_buffer.hpp>
 
-#include <boost/buffers/type_traits.hpp>
+#include <boost/buffers/dynamic_buffer.hpp>
 #include <boost/static_assert.hpp>
 #include "test_buffers.hpp"
